Added unicode title support
Added support for unicode characters in track titles. Non-English Unicode letters are transliterated to ASCII for the track internal names.

Note that the tool does NOT try to translate the titles, it just uses what the transliteration library returns as the track's internal name. The tool just needs a unique identifier for each track, it doesn't care how badly translated they might be.